feat: do not autocomplete already used cmd flags
When autocompleting commands, completer would suggest same flag again.
E.g. for command Show if user provided -s and again press TAB on '-' it
would again suggest -s

Changes in this commit will filter out already existing flags so in
previous example -s would not be suggested.

Additionally, if a flag has a short and long version, then if short is
already existing long won't be suggested anymore (and vice versa).
E.g. in command Show if user provided -t (totp flag) and then press TAB
on '-' or '--' or '--t' it will not suggest '--totp' since it's the same
flag as '-t'